projects/Ip_tokenization-contracts/pyproject.toml
[tool.poetry]
name = "Ip_tokenization-contracts"
version = "0.1.0"
description = "Algorand smart contracts"
authors = ["mansi754 <mansimladdha@gmail.com>"]
readme = "README.md"
packages = [
    { include = "smart_contracts" }
]

[tool.poetry.dependencies]
python = "^3.12"
algokit-utils = "^4.0.0"
python-dotenv = "^1.0.0"
pyteal = "^0.26.0"
beaker = "^1.1.0"

[tool.poetry.group.dev.dependencies]
black = {version = "^23.3.0", extras = ["d"]}
mypy = "^1.3.0"
pip-audit = "^2.5.4"
pytest = "^7.3.1"
pytest-cov = "^4.0.0"
python-semantic-release = "^7.33.2"
ruff = "^0.0.270"
setuptools = "^67.8.0"
types-setuptools = "^67.8.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

